matcher 改造、博客网站项目初始化

+++ core/src/main/java/com/commons/core/enums/CommonStatusEnum.java
@@ -4,7 +4,7 @@ import lombok.Getter;
 
 /**
  * Description:
- *      通用状态码
+ * 通用状态码
  * Project: commons
  * Author: cccsp
  * Create Time:2022/9/19 16:08
@@ -15,33 +15,33 @@ public enum CommonStatusEnum {
     /**
      * 是
      */
-    TRUE(1, "是"),
+    TRUE((byte) 1, "是"),
 
     /**
      * 否
      */
-    FALSE(0, "否");
+    FALSE((byte) 0, "否");
 
     /**
      * 状态码
      */
-    private final Integer statusCode;
+    private final byte statusCode;
 
     /**
      * 描述
      */
     private final String desc;
 
-    CommonStatusEnum(Integer statusCode, String desc) {
+    CommonStatusEnum(byte statusCode, String desc) {
         this.statusCode = statusCode;
         this.desc = desc;
     }
 
-    public static Integer getTrue() {
+    public static byte getTrue() {
         return TRUE.statusCode;
     }
 
-    public static Integer getFalse() {
+    public static byte getFalse() {
         return FALSE.statusCode;
     }
 }
